Extract retained to separate project + clean up dependencies
