+++ circuit-retained/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuit-retained
+POM_NAME=Circuit (retained)
+POM_DESCRIPTION=Circuit (Retained)
