+++ circuit-retained/src/main/kotlin/com/slack/circuit/retained/CanRetainChecker.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2022 Slack Technologies, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.slack.circuit.retained
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ProvidedValue
+import androidx.compose.runtime.staticCompositionLocalOf
+
+class CanRetainChecker(private val shouldRetain: () -> Boolean) {
+  fun shouldRetain() = shouldRetain.invoke()
+}
+
+object LocalCanRetainCheckerOwner {
+  private val LocalCanRetainChecker = staticCompositionLocalOf { CanRetainChecker { false } }
+
+  /**
+   * Returns current composition local value for the owner or errors if one has not been provided.
+   */
+  val current: CanRetainChecker
+    @Composable get() = LocalCanRetainChecker.current
+
+  /**
+   * Associates a [LocalCanRetainChecker] key to a value in a call to [CompositionLocalProvider].
+   */
+  infix fun provides(shouldRetain: () -> Boolean): ProvidedValue<CanRetainChecker> {
+    return LocalCanRetainChecker.provides(CanRetainChecker(shouldRetain))
+  }
+}
