+++ circuit/src/main/kotlin/com/slack/circuit/CircuitProvider.kt
@@ -26,6 +26,7 @@ import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.slack.circuit.retained.Continuity
+import com.slack.circuit.retained.LocalCanRetainCheckerOwner
 import com.slack.circuit.retained.LocalRetainedStateRegistry
 
 /** Provides the given [circuit] as a [CompositionLocal] to all composables within [content]. */
@@ -54,27 +55,6 @@ object LocalCircuitOwner {
   /** Associates a [LocalCircuit] key to a value in a call to [CompositionLocalProvider]. */
   infix fun provides(circuit: Circuit): ProvidedValue<Circuit?> {
     return LocalCircuit.provides(circuit)
-  }
-}
-
-class CanRetainChecker(private val shouldRetain: () -> Boolean) {
-  fun shouldRetain() = shouldRetain.invoke()
-}
-
-object LocalCanRetainCheckerOwner {
-  private val LocalCanRetainChecker = staticCompositionLocalOf { CanRetainChecker { false } }
-
-  /**
-   * Returns current composition local value for the owner or errors if one has not been provided.
-   */
-  val current: CanRetainChecker
-    @Composable get() = LocalCanRetainChecker.current
-
-  /**
-   * Associates a [LocalCanRetainChecker] key to a value in a call to [CompositionLocalProvider].
-   */
-  infix fun provides(shouldRetain: () -> Boolean): ProvidedValue<CanRetainChecker> {
-    return LocalCanRetainChecker.provides(CanRetainChecker(shouldRetain))
   }
 }
